Move rms_norm kernel into its own normalization subfolder.
Keep layout consistent with group_norm_grad and update the test include
plus skill path references.
